+++ ftplugin/markdown.lua
@@ -5,4 +5,4 @@ vim.b.slime_cell_delimiter = "```"
 vim.wo.wrap = true
 vim.wo.linebreak = true
 vim.wo.breakindent = true
-vim.wo.showbreak = '|'
+vim.wo.showbreak = "|"