+++ init.lua
@@ -1,8 +1,5 @@
-require 'config.global'
-require 'config.lazy'
-require 'config.autocommands'
-require 'config.keymap'
-require 'config.color'
-
--- require'config.experimental'
-
+require("config.global")
+require("config.lazy")
+require("config.autocommands")
+require("config.keymap")
+require("config.color")