+++ lua/config/global.lua
@@ -2,17 +2,16 @@
 vim.opt.termguicolors = true
 
 -- more opinionated
-vim.opt.number = true                  -- show linenumbers
-vim.opt.mouse = 'a'                    -- enable mouse
+vim.opt.number = true -- show linenumbers
+vim.opt.mouse = "a" -- enable mouse
 vim.opt.mousefocus = true
-vim.opt.clipboard:append 'unnamedplus' -- use system clipboard
-
+vim.opt.clipboard:append("unnamedplus") -- use system clipboard
 
 vim.opt.timeoutlen = 400 -- until which-key pops up
 vim.opt.updatetime = 250 -- for autocommands and hovers
 
 -- don't ask about existing swap files
-vim.opt.shortmess:append 'A'
+vim.opt.shortmess:append("A")
 
 -- use spaces as tabs
 local tabsize = 2
@@ -36,18 +35,18 @@ vim.opt.breakindent = true
 vim.opt.signcolumn = "yes:1"
 
 -- how to show autocomplete menu
-vim.opt.completeopt = 'menuone,noinsert'
+vim.opt.completeopt = "menuone,noinsert"
 
 -- add folds with treesitter grammar
--- vim.opt.foldmethod = "expr"
--- vim.wo.foldexpr = 'v:lua.vim.treesitter.foldexpr()'
--- -- but open all by default
--- vim.opt.foldlevel = 99
+vim.opt.foldmethod = "expr"
+vim.wo.foldexpr = "v:lua.vim.treesitter.foldexpr()"
+-- but open all by default
+vim.opt.foldlevel = 99
 
 -- global statusline
 vim.opt.laststatus = 3
 
-vim.cmd [[
+vim.cmd([[
 let g:currentmode={
        \ 'n'  : '%#String# NORMAL ',
        \ 'v'  : '%#Search# VISUAL ',
@@ -59,8 +58,8 @@ let g:currentmode={
        \ 'c'  : '%#CurSearch# Command ',
        \ 't'  : '%#ModeMsg# TERM ',
        \}
-]]
-vim.opt.statusline = '%{%g:currentmode[mode()]%} %* %t | %y | %* %= c:%c l:%l/%L %p%% 🦦 '
+]])
+vim.opt.statusline = "%{%g:currentmode[mode()]%} %* %t | %y | %* %= c:%c l:%l/%L %p%% 🦦 "
 
 -- split right and below by default
 vim.opt.splitright = true
@@ -70,17 +69,16 @@ vim.opt.splitbelow = true
 vim.opt.showtabline = 1
 
 --windowline
-vim.opt.winbar = '%f'
+vim.opt.winbar = "%f"
 
 --don't continue comments automagically
-vim.opt.formatoptions:remove({ 'c', 'r', 'o' })
+vim.opt.formatoptions:remove({ "c", "r", "o" })
 
 -- hide cmdline when not used
-vim.opt.cmdheight = 1
+vim.opt.cmdheight = 0
 
 -- scroll before end of window
 vim.opt.scrolloff = 5
 
--- (don't) replace certain elements with prettier ones
+-- (don't == 0) replace certain elements with prettier ones
 vim.opt.conceallevel = 0
-